feat: kernel module build system

+++ scripts/hdd/make-hdd.sh
@@ -14,14 +14,30 @@ PART2_END=$(( $SECTOR_END - 1 ))
 PART1_START=2048
 PART1_END=$(( $PART2_START - 1 ))
 
-echo CREATE $(basename $HDD) "($HDD_SIZE_MB""MB)"
+if [ $(uname -s) = 'Darwin' ]; then
+    MAX_RETRY=10
+    ATTEMPT=1
+    until docker ps > /dev/null 2> /dev/null || [ $MAX_RETRY -le $ATTEMPT ] ; do
+        echo TRY-OPEN docker "($ATTEMPT / $MAX_RETRY)"
+        open -g -a docker
+        ATTEMPT=$(( $ATTEMPT + 1 ))
+        sleep 5
+    done
+fi
+
+if ! docker ps > /dev/null 2> /dev/null; then
+    echo error: docker daemon is not running.
+    exit 1
+fi
+
+echo QEMU-IMG $(basename $HDD) "($HDD_SIZE_MB""MB)"
 qemu-img create -q -f qcow2 $HDD $HDD_SIZE_MB"M"
 qemu-nbd --persistent $HDD &
 NBD_SERVER=$!
-trap "kill $NBD_SERVER" EXIT 
+trap "kill $NBD_SERVER" EXIT  
 
-echo DOCKER-RUN bkahlert/libguestfs:edge
-docker run --rm -i --add-host=host.docker.internal:host-gateway bkahlert/libguestfs:edge guestfish << EOF
+echo DOCKER-RUN guestfish
+docker run -v$(pwd)/target:/target --rm -i --add-host=host.docker.internal:host-gateway bkahlert/libguestfs:edge guestfish << EOF
 add '' protocol:nbd server:host.docker.internal
 run
 part-init /dev/sda mbr
@@ -29,4 +45,6 @@ part-add /dev/sda p $PART1_START $PART1_END
 part-add /dev/sda p $PART2_START $PART2_END
 mkfs ext2 /dev/sda1 blocksize:1024
 mkfs ext2 /dev/sda2
+mount /dev/sda1 /
+upload /target/i686-unknown-none-elf/debug/hello.ko /hello.ko
 EOF
